Improvements to UKPhoneNumberParser
Store separate lists of AreaCodeInfo per phone number kind to reduce the search cost when parsing a phone number, e.g. avoids searching mobile ranges when parsing a geo number.

Generate the 5 digit area code lookup based upon the contents of the data file and cater for regions that have a "child range" which is a subset of a parent range.