Reduce array allocations
Each formatter is a thread safe singleton, however every CountryInfo instance is creating an instance array of them. Since the only operation is finding a formatter from the array it is safe to share a static collection across all CountryInfo instances. This will need to be revisited if the option to allow external control of formatters is added in the future.